rename MainActivity to be DashboardActivity
