+++ app/src/main/AndroidManifest.xml
@@ -55,7 +55,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.GodTools">
         <activity
-            android:name="org.keynote.godtools.android.activity.MainActivity"
+            android:name=".ui.dashboard.DashboardActivity"
             android:exported="true"
             android:launchMode="singleTask"
             android:screenOrientation="@integer/default_screen_orientation">
@@ -90,18 +90,18 @@
         <activity
             android:name=".ui.about.AboutActivity"
             android:label="@string/title_about"
-            android:parentActivityName="org.keynote.godtools.android.activity.MainActivity"
+            android:parentActivityName=".ui.dashboard.DashboardActivity"
             android:screenOrientation="@integer/default_screen_orientation" />
 
         <activity
             android:name=".ui.tooldetails.ToolDetailsActivity"
-            android:parentActivityName="org.keynote.godtools.android.activity.MainActivity"
+            android:parentActivityName=".ui.dashboard.DashboardActivity"
             android:screenOrientation="@integer/default_screen_orientation" />
 
         <activity
             android:name=".ui.languages.LanguageSettingsActivity"
             android:label="@string/title_language_settings"
-            android:parentActivityName="org.keynote.godtools.android.activity.MainActivity"
+            android:parentActivityName=".ui.dashboard.DashboardActivity"
             android:screenOrientation="@integer/default_screen_orientation" />
 
         <activity
@@ -111,20 +111,20 @@
 
         <activity
             android:name=".tool.cyoa.ui.CyoaActivity"
-            android:parentActivityName="org.keynote.godtools.android.activity.MainActivity" />
+            android:parentActivityName=".ui.dashboard.DashboardActivity" />
 
         <activity
             android:name=".tract.activity.TractActivity"
-            android:parentActivityName="org.keynote.godtools.android.activity.MainActivity" />
+            android:parentActivityName=".ui.dashboard.DashboardActivity" />
 
         <activity
             android:name=".tutorial.activity.TutorialActivity"
-            android:parentActivityName="org.keynote.godtools.android.activity.MainActivity" />
+            android:parentActivityName=".ui.dashboard.DashboardActivity" />
 
         <activity
             android:name=".ui.profile.ProfileActivity"
             android:label="@string/profile_title"
-            android:parentActivityName="org.keynote.godtools.android.activity.MainActivity" />
+            android:parentActivityName=".ui.dashboard.DashboardActivity" />
 
         <provider
             android:name="androidx.startup.InitializationProvider"
