+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/DashboardAppsFlyerDeepLinkResolver.kt
@@ -6,11 +6,11 @@ import android.net.Uri
 import org.cru.godtools.analytics.appsflyer.AppsFlyerDeepLinkResolver
 import org.cru.godtools.base.ui.createDashboardIntent
 import org.cru.godtools.base.ui.dashboard.Page
-import org.keynote.godtools.android.activity.MainActivity
 
 internal object DashboardAppsFlyerDeepLinkResolver : AppsFlyerDeepLinkResolver {
     override fun resolve(context: Context, uri: Uri?, data: Map<String, String?>) = when {
-        uri?.isDashboardLessonsDeepLink() == true -> Intent(Intent.ACTION_VIEW, uri, context, MainActivity::class.java)
+        uri?.isDashboardLessonsDeepLink() == true ->
+            Intent(Intent.ACTION_VIEW, uri, context, DashboardActivity::class.java)
         else -> null
     }
